+++ tests/cover/test_charmap.py
@@ -111,14 +111,29 @@ def test_recreate_charmap():
 
 
 def test_union_empty():
-    assert cm._union_interval_lists([], [[1, 2]]) == [[1, 2]]
-    assert cm._union_interval_lists([[1, 2]], []) == [[1, 2]]
+    assert cm._union_intervals([], []) == ()
+    assert cm._union_intervals([], [[1, 2]]) == ((1, 2),)
+    assert cm._union_intervals([[1, 2]], []) == ((1, 2),)
+
+
+def test_union_handles_totally_overlapped_gap():
+    #   < xx  >  Imagine the intervals x and y as bit strings.
+    # | <yy yy>  The bit at position n is set if n falls inside that interval.
+    # = <zzzzz>  In this model _union_intervals() performs bit-wise or.
+    assert cm._union_intervals([[2, 3]], [[1, 2], [4, 5]]) == ((1, 5),)
+
+
+def test_union_handles_partially_overlapped_gap():
+    #   <  x  >  Imagine the intervals x and y as bit strings.
+    # | <yy  y>  The bit at position n is set if n falls inside that interval.
+    # = <zzz z>  In this model _union_intervals() performs bit-wise or.
+    assert cm._union_intervals([[3, 3]], [[1, 2], [5, 5]]) == ((1, 3), (5, 5))
 
 
 def test_successive_union():
     x = []
     for v in cm.charmap().values():
-        x = cm._union_interval_lists(x, v)
+        x = cm._union_intervals(x, v)
     assert x == ((0, sys.maxunicode),)